+++ crates/wasi-nn/examples/stable-diffusion/src/schedulers/lms_discrete.rs
@@ -6,9 +6,9 @@ use ndarray::linspace;
 #[derive(Debug, Clone)]
 pub struct LMSDiscreteSchedulerConfig {
     /// The value of beta at the beginning of training.
-    pub beta_start: f64,
+    pub beta_start: f32,
     /// The value of beta at the end of training.
-    pub beta_end: f64,
+    pub beta_end: f32,
     /// How beta evolved during training.
     pub beta_schedule: BetaSchedule,
     /// number of diffusion steps used to train the model.
@@ -34,34 +34,34 @@ impl Default for LMSDiscreteSchedulerConfig {
 }
 
 pub struct LMSDiscreteScheduler {
-    timesteps: Vec<f64>,
-    sigmas: Vec<f64>,
-    init_noise_sigma: f64,
-    derivatives: Vec<Vec<f64>>,
+    timesteps: Vec<f32>,
+    sigmas: Vec<f32>,
+    init_noise_sigma: f32,
+    derivatives: Vec<Vec<f32>>,
     pub config: LMSDiscreteSchedulerConfig,
 }
 
 impl LMSDiscreteScheduler {
     pub fn new(inference_steps: usize, config: LMSDiscreteSchedulerConfig) -> Self {
-        let betas: Vec<f64> = match config.beta_schedule {
+        let betas: Vec<f32> = match config.beta_schedule {
             BetaSchedule::ScaledLinear => ndarray::linspace(
                 config.beta_start.sqrt(),
                 config.beta_end.sqrt(),
                 config.train_timesteps,
             )
             .map(|x| x.powi(2))
-            .collect::<Vec<f64>>(),
+            .collect::<Vec<f32>>(),
             BetaSchedule::Linear => {
                 ndarray::linspace(config.beta_start, config.beta_end, config.train_timesteps)
-                    .collect::<Vec<f64>>()
+                    .collect::<Vec<f32>>()
             }
             _ => unimplemented!(
                 "LMSDiscreteScheduler only implements linear and scaled_linear betas."
             ),
         };
 
-        let alphas: Vec<f64> = betas.iter().map(|&x| 1. - x).collect();
-        let alphas_cumprod: Vec<f64> = alphas
+        let alphas: Vec<f32> = betas.iter().map(|&x| 1. - x).collect();
+        let alphas_cumprod: Vec<f32> = alphas
             .iter()
             .scan(1.0, |state, &x| {
                 *state *= x;
@@ -70,26 +70,26 @@ impl LMSDiscreteScheduler {
             .collect();
 
         let timesteps = ndarray::linspace(
-            (config.train_timesteps - 1) as f64,
+            (config.train_timesteps - 1) as f32,
             0.,
             inference_steps as usize,
         )
-        .collect::<Vec<f64>>();
+        .collect::<Vec<f32>>();
 
         let sigmas = alphas_cumprod
             .iter()
             .map(|&x| ((1. - x) / x).sqrt())
-            .collect::<Vec<f64>>();
+            .collect::<Vec<f32>>();
         let sigmas = Self::interp(
             timesteps.as_slice(), // x-coordinates at which to evaluate the interpolated values
-            ndarray::range(0., sigmas.len() as f64, 1.)
-                .collect::<Vec<f64>>()
+            ndarray::range(0., sigmas.len() as f32, 1.)
+                .collect::<Vec<f32>>()
                 .as_slice(),
             sigmas.as_slice(),
         );
 
         // standard deviation of the initial noise distribution
-        let init_noise_sigma: f64 = sigmas.iter().fold(f64::MIN, |a, &b| a.max(b));
+        let init_noise_sigma: f32 = sigmas.iter().fold(f32::MIN, |a, &b| a.max(b));
 
         Self {
             timesteps: timesteps.try_into().unwrap(),
@@ -100,12 +100,12 @@ impl LMSDiscreteScheduler {
         }
     }
 
-    pub fn timesteps(&self) -> &[f64] {
+    pub fn timesteps(&self) -> &[f32] {
         self.timesteps.as_slice()
     }
 
     /// Scales the denoising model input by `(sigma^2 + 1)^0.5` to match the K-LMS algorithm.
-    pub fn scale_model_input(&self, sample: Vec<f64>, timestep: f64) -> Vec<f32> {
+    pub fn scale_model_input(&self, sample: Vec<f32>, timestep: f32) -> Vec<f32> {
         let step_index = self.timesteps.iter().position(|&t| t == timestep).unwrap();
         let sigma = self.sigmas[step_index];
 
@@ -120,9 +120,9 @@ impl LMSDiscreteScheduler {
     /// points, mimicking np.interp().
     ///
     /// Based on https://github.com/pytorch/pytorch/issues/50334#issuecomment-1000917964
-    fn interp(timesteps: &[f64], range: &[f64], sigmas: &[f64]) -> Vec<f64> {
+    fn interp(timesteps: &[f32], range: &[f32], sigmas: &[f32]) -> Vec<f32> {
         let sz = timesteps.len();
-        let mut result: Vec<f64> = vec![0.; sz];
+        let mut result: Vec<f32> = vec![0.; sz];
 
         for i in 1..sz {
             let index = range.binary_search_by(|a| a.partial_cmp(&timesteps[i]).unwrap());
@@ -146,8 +146,8 @@ impl LMSDiscreteScheduler {
     }
 
     /// Compute a linear multistep coefficient
-    fn get_lms_coefficient(&mut self, order: usize, t: usize, current_order: usize) -> f64 {
-        let lms_derivative = |tau| -> f64 {
+    fn get_lms_coefficient(&mut self, order: usize, t: usize, current_order: usize) -> f32 {
+        let lms_derivative = |tau| -> f32 {
             let mut prod = 1.0;
             for k in 0..order {
                 if current_order == k {
@@ -169,12 +169,12 @@ impl LMSDiscreteScheduler {
         integration_out.integral
     }
 
-    pub fn step(&mut self, model_output: &[f64], timestep: f64, sample: &[f64]) -> Vec<f64> {
+    pub fn step(&mut self, model_output: &[f32], timestep: f32, sample: &[f32]) -> Vec<f32> {
         let step_index = self.timesteps.iter().position(|&t| t == timestep).unwrap();
         let sigma = self.sigmas[step_index];
 
         // 1. compute predicted original sample (x_0) from sigma-scaled predicted noise
-        let pred_original_sample: Vec<f64> = match self.config.prediction_type {
+        let pred_original_sample: Vec<f32> = match self.config.prediction_type {
             PredictionType::Epsilon => {
                 let mut result = vec![0.; sample.len()];
                 for i in 0..sample.len() {
@@ -194,7 +194,7 @@ impl LMSDiscreteScheduler {
         };
 
         // 2. Convert to an ODE derivative
-        let mut derivative: Vec<f64> = vec![0.0; sample.len()];
+        let mut derivative: Vec<f32> = vec![0.0; sample.len()];
         for i in 0..sample.len() {
             derivative[i] = (sample[i] - pred_original_sample[i]) / sigma;
         }
@@ -212,25 +212,25 @@ impl LMSDiscreteScheduler {
 
         // 4. compute previous sample based on the derivatives path
         // https://github.com/huggingface/diffusers/blob/769f0be8fb41daca9f3cbcffcfd0dbf01cc194b8/src/diffusers/schedulers/scheduling_lms_discrete.py#L243-L245
-        let deriv_prod: Vec<Vec<f64>> = lms_coeffs
+        let deriv_prod: Vec<Vec<f32>> = lms_coeffs
             .iter()
             .zip(self.derivatives.iter().rev())
             .map(|(coeff, derivative)| {
-                let mut result: Vec<f64> = vec![0.; sample.len()];
+                let mut result: Vec<f32> = vec![0.; sample.len()];
                 for i in 0..derivative.len() {
                     result[i] = derivative[i] * coeff
                 }
                 result
             })
             .collect();
-        let mut deriv_sum: Vec<f64> = vec![0.; sample.len()];
+        let mut deriv_sum: Vec<f32> = vec![0.; sample.len()];
         for i in 0..deriv_prod.len() {
             for j in 0..deriv_prod[i].len() {
                 deriv_sum[j] += deriv_prod[i][j]
             }
         }
 
-        let mut result: Vec<f64> = vec![0.; sample.len()];
+        let mut result: Vec<f32> = vec![0.; sample.len()];
         for i in 0..sample.len() {
             result[i] = deriv_sum[i] + sample[i]
         }
@@ -238,16 +238,16 @@ impl LMSDiscreteScheduler {
         result
     }
 
-    pub fn init_noise_sigma(&self) -> f64 {
+    pub fn init_noise_sigma(&self) -> f32 {
         self.init_noise_sigma
     }
 
-    pub fn add_noise(&self, original_samples: &[f64], noise: &[f64], timestep: f64) -> Vec<f64> {
+    pub fn add_noise(&self, original_samples: &[f32], noise: &[f32], timestep: f32) -> Vec<f32> {
         let step_index = self.timesteps.iter().position(|&t| t == timestep).unwrap();
         let sigma = self.sigmas[step_index];
 
         // noisy samples
-        let mut samples: Vec<f64> = vec![0.; original_samples.len()];
+        let mut samples: Vec<f32> = vec![0.; original_samples.len()];
         for i in 0..original_samples.len() {
             samples[i] = original_samples[i] + noise[i] * sigma
         }
